Replace f64 with Decimal; add Symbol.margin_asset

+++ src/account.rs
@@ -1,3 +1,5 @@
+use rust_decimal::Decimal;
+
 use crate::client::*;
 use crate::errors::*;
 use crate::rest_model::*;
@@ -30,15 +32,15 @@ pub struct OrderRequest {
     #[serde(rename = "type")]
     pub order_type: OrderType,
     pub time_in_force: Option<TimeInForce>,
-    pub quantity: Option<f64>,
-    pub quote_order_qty: Option<f64>,
-    pub price: Option<f64>,
+    pub quantity: Option<Decimal>,
+    pub quote_order_qty: Option<Decimal>,
+    pub price: Option<Decimal>,
     /// A unique id for the order, automatically generated if not sent.
     pub new_client_order_id: Option<String>,
     /// Used with stop loss, stop loss limit, take profit and take profit limit order types.
-    pub stop_price: Option<f64>,
+    pub stop_price: Option<Decimal>,
     /// Used with limit, stop loss limit and take profit limit to create an iceberg order.
-    pub iceberg_qty: Option<f64>,
+    pub iceberg_qty: Option<Decimal>,
     /// Set the response json, market and limit default to full others to ack.
     pub new_order_resp_type: Option<OrderResponse>,
     /// Cannot be greater than 60000
