+++ src/ws_model.rs
@@ -1,4 +1,6 @@
-use crate::rest_model::{string_or_float, Asks, Bids, OrderBook, OrderSide, OrderStatus, OrderType, TimeInForce};
+use rust_decimal::Decimal;
+
+use crate::rest_model::{as_decimal, Asks, Bids, OrderBook, OrderSide, OrderStatus, OrderType, TimeInForce};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(tag = "e")]
@@ -195,34 +197,34 @@ pub struct Kline {
     pub first_trade_id: i64,
     #[serde(rename = "L")]
     pub last_trade_id: i64,
-    #[serde(with = "string_or_float")]
+    #[serde(with = "as_decimal")]
     #[serde(rename = "o")]
-    pub open: f64,
-    #[serde(with = "string_or_float")]
+    pub open: Decimal,
+    #[serde(with = "as_decimal")]
     #[serde(rename = "c")]
-    pub close: f64,
-    #[serde(with = "string_or_float")]
+    pub close: Decimal,
+    #[serde(with = "as_decimal")]
     #[serde(rename = "h")]
-    pub high: f64,
-    #[serde(with = "string_or_float")]
+    pub high: Decimal,
+    #[serde(with = "as_decimal")]
     #[serde(rename = "l")]
-    pub low: f64,
-    #[serde(with = "string_or_float")]
+    pub low: Decimal,
+    #[serde(with = "as_decimal")]
     #[serde(rename = "v")]
-    pub volume: f64,
+    pub volume: Decimal,
     #[serde(rename = "n")]
     pub number_of_trades: i64,
     #[serde(rename = "x")]
     pub is_final_bar: bool,
-    #[serde(with = "string_or_float")]
+    #[serde(with = "as_decimal")]
     #[serde(rename = "q")]
-    pub quote_volume: f64,
-    #[serde(with = "string_or_float")]
+    pub quote_volume: Decimal,
+    #[serde(with = "as_decimal")]
     #[serde(rename = "V")]
-    pub active_buy_volume: f64,
-    #[serde(with = "string_or_float")]
+    pub active_buy_volume: Decimal,
+    #[serde(with = "as_decimal")]
     #[serde(rename = "Q")]
-    pub active_volume_buy_quote: f64,
+    pub active_volume_buy_quote: Decimal,
     #[serde(skip, rename = "B")]
     pub ignore_me: String,
 }
@@ -253,17 +255,17 @@ pub struct BookTickerEvent {
     #[serde(rename = "s")]
     pub symbol: String,
 
-    #[serde(rename = "b", with = "string_or_float")]
-    pub best_bid: f64,
+    #[serde(rename = "b", with = "as_decimal")]
+    pub best_bid: Decimal,
 
-    #[serde(rename = "B", with = "string_or_float")]
-    pub best_bid_qty: f64,
+    #[serde(rename = "B", with = "as_decimal")]
+    pub best_bid_qty: Decimal,
 
-    #[serde(rename = "a", with = "string_or_float")]
-    pub best_ask: f64,
+    #[serde(rename = "a", with = "as_decimal")]
+    pub best_ask: Decimal,
 
-    #[serde(rename = "A", with = "string_or_float")]
-    pub best_ask_qty: f64,
+    #[serde(rename = "A", with = "as_decimal")]
+    pub best_ask_qty: Decimal,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -345,11 +347,11 @@ pub struct EventBalance {
     #[serde(rename = "a")]
     pub asset: String,
     #[serde(rename = "f")]
-    #[serde(with = "string_or_float")]
-    pub free: f64,
+    #[serde(with = "as_decimal")]
+    pub free: Decimal,
     #[serde(rename = "l")]
-    #[serde(with = "string_or_float")]
-    pub locked: f64,
+    #[serde(with = "as_decimal")]
+    pub locked: Decimal,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
@@ -362,8 +364,8 @@ pub struct BalanceUpdate {
     pub asset: String,
 
     #[serde(rename = "d")]
-    #[serde(with = "string_or_float")]
-    pub delta: f64,
+    #[serde(with = "as_decimal")]
+    pub delta: Decimal,
 
     #[serde(alias = "T")]
     pub clear_time: u64,
@@ -385,17 +387,17 @@ pub struct OrderUpdate {
     #[serde(rename = "f")]
     pub time_in_force: TimeInForce,
     #[serde(rename = "q")]
-    #[serde(with = "string_or_float")]
-    pub qty: f64,
+    #[serde(with = "as_decimal")]
+    pub qty: Decimal,
     #[serde(rename = "p")]
-    #[serde(with = "string_or_float")]
-    pub price: f64,
+    #[serde(with = "as_decimal")]
+    pub price: Decimal,
     #[serde(rename = "P")]
-    #[serde(with = "string_or_float")]
-    pub stop_price: f64,
+    #[serde(with = "as_decimal")]
+    pub stop_price: Decimal,
     #[serde(rename = "F")]
-    #[serde(with = "string_or_float")]
-    pub iceberg_qty: f64,
+    #[serde(with = "as_decimal")]
+    pub iceberg_qty: Decimal,
     #[serde(rename = "g")]
     pub order_list_id: i64,
     #[serde(rename = "C")]
@@ -409,17 +411,17 @@ pub struct OrderUpdate {
     #[serde(rename = "i")]
     pub order_id: u64,
     #[serde(rename = "l")]
-    #[serde(with = "string_or_float")]
-    pub qty_last_executed: f64,
+    #[serde(with = "as_decimal")]
+    pub qty_last_executed: Decimal,
     #[serde(rename = "z")]
-    #[serde(with = "string_or_float")]
-    pub cumulative_filled_qty: f64,
+    #[serde(with = "as_decimal")]
+    pub cumulative_filled_qty: Decimal,
     #[serde(rename = "L")]
-    #[serde(with = "string_or_float")]
-    pub last_executed_price: f64,
+    #[serde(with = "as_decimal")]
+    pub last_executed_price: Decimal,
     #[serde(rename = "n")]
-    #[serde(with = "string_or_float")]
-    pub commission: f64,
+    #[serde(with = "as_decimal")]
+    pub commission: Decimal,
     #[serde(rename = "N")]
     pub commission_asset: Option<String>,
     #[serde(rename = "T")]
@@ -437,15 +439,15 @@ pub struct OrderUpdate {
     #[serde(rename = "O")]
     pub order_creation_time: u64,
     #[serde(rename = "Z")]
-    #[serde(with = "string_or_float")]
-    pub cumulative_quote_asset_transacted_qty: f64,
+    #[serde(with = "as_decimal")]
+    pub cumulative_quote_asset_transacted_qty: Decimal,
     /// (i.e. lastPrice * lastQty)
     #[serde(rename = "Y")]
-    #[serde(with = "string_or_float")]
-    pub last_quote_asset_transacted_qty: f64,
+    #[serde(with = "as_decimal")]
+    pub last_quote_asset_transacted_qty: Decimal,
     #[serde(rename = "Q")]
-    #[serde(with = "string_or_float")]
-    pub quote_order_qty: f64,
+    #[serde(with = "as_decimal")]
+    pub quote_order_qty: Decimal,
 }
 
 /// For OCO Events
